fix(e2e): recognize dashboard forward recovery
