+++ test/e2e/support/dashboard-remote-bind-env.test.ts
@@ -4,7 +4,10 @@
 import { describe, expect, it } from "vitest";
 
 import { buildAvailabilityProbeEnv } from "../fixtures/availability-env.ts";
-import { buildDashboardRemoteBindEnv } from "../live/dashboard-remote-bind-env.ts";
+import {
+  buildDashboardRemoteBindEnv,
+  dashboardRemoteBindConnectStarted,
+} from "../live/dashboard-remote-bind-env.ts";
 
 describe("dashboard remote-bind E2E environment", () => {
   it("prepares remote dashboard exposure during install and onboarding", () => {
@@ -27,5 +30,33 @@ describe("dashboard remote-bind E2E environment", () => {
     });
 
     expect(env.NEMOCLAW_DASHBOARD_BIND).toBe("0.0.0.0");
+  });
+
+  it("accepts recovery proof when connect has no numeric exit code", () => {
+    expect(
+      dashboardRemoteBindConnectStarted(
+        {
+          exitCode: null,
+          stdout: "\u001B[32m✓\u001B[0m Dashboard port forward re-established.\n",
+          stderr: "client_loop: send disconnect: Broken pipe\n",
+        },
+        "e2e-dashboard-remote-bind",
+        "18789",
+      ),
+    ).toBe(true);
+  });
+
+  it("rejects a connect result with no numeric exit code and no forward proof", () => {
+    expect(
+      dashboardRemoteBindConnectStarted(
+        {
+          exitCode: null,
+          stdout: "Connecting to sandbox...\n",
+          stderr: "",
+        },
+        "e2e-dashboard-remote-bind",
+        "18789",
+      ),
+    ).toBe(false);
   });
 });
